requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
lxml==5.2.1
lxpy==1.3.9
PyExecJS==1.5.1
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
